merge code

+++ jaxp/test/javax/xml/jaxp/unittest/parsers/Bug6341770.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2015, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2016, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -23,29 +23,45 @@
 
 package parsers;
 
+import static jaxp.library.JAXPTestUtilities.tryRunWithTmpPermission;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.PrintWriter;
+import java.nio.file.Paths;
+import java.util.PropertyPermission;
 
 import javax.xml.parsers.SAXParserFactory;
 
 import org.testng.Assert;
+import org.testng.annotations.Listeners;
 import org.testng.annotations.Test;
 import org.xml.sax.Attributes;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.DefaultHandler;
 
 /*
+ * @test
  * @bug 6341770
+ * @library /javax/xml/jaxp/libs /javax/xml/jaxp/unittest
+ * @run testng/othervm -DrunSecMngr=true parsers.Bug6341770
+ * @run testng/othervm parsers.Bug6341770
  * @summary Test external entity linked to non-ASCII base URL.
  */
+@Listeners({jaxp.library.FilePolicy.class})
 public class Bug6341770 {
 
     // naming a file "aux" would fail on windows.
     @Test
     public void testNonAsciiURI() {
+        if (!isNonAsciiSupported()) {
+            // @bug 8167478
+            // if it doesn't support non-ascii, the following test is invalid even if test is passed.
+            System.out.println("Current environment doesn't support non-ascii, exit the test.");
+            return;
+        }
         try {
-            File dir = File.createTempFile("sko\u0159ice", null);
+            File dir = new File(ALPHA);
             dir.delete();
             dir.mkdir();
             File main = new File(dir, "main.xml");
@@ -60,15 +76,31 @@ public class Bug6341770 {
             w.flush();
             w.close();
             System.out.println("Parsing: " + main);
-            SAXParserFactory.newInstance().newSAXParser().parse(main, new DefaultHandler() {
-                public void startElement(String uri, String localname, String qname, Attributes attr) throws SAXException {
-                    System.out.println("encountered <" + qname + ">");
-                }
-            });
+            tryRunWithTmpPermission(
+                    () -> SAXParserFactory.newInstance().newSAXParser().parse(main, new DefaultHandler() {
+                        public void startElement(String uri, String localname, String qname, Attributes attr)
+                                throws SAXException {
+                            System.out.println("encountered <" + qname + ">");
+                        }
+                    }), new PropertyPermission("user.dir", "read"));
         } catch (Exception e) {
             e.printStackTrace();
             Assert.fail("Exception: " + e.getMessage());
         }
         System.out.println("OK.");
     }
+
+    private boolean isNonAsciiSupported() {
+        // Use Paths.get method to test if the path is valid in current environment
+        try {
+            Paths.get(ALPHA);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    // Select alpha because it's a very common non-ascii character in different charsets.
+    // That this test can run in as many as possible environments if it's possible.
+    private static final String ALPHA = "\u03b1";
 }
